+++ src/bll/product.go
@@ -76,11 +76,15 @@ func (a *Product) Offline(ctx context.Context, product string) (*tpl.BoolRes, er
 
 // Delete 逻辑删除产品
 func (a *Product) Delete(ctx context.Context, product string) (*tpl.BoolRes, error) {
-	err := daos.UrbsAcAcl.DeleteByObject(ctx, product)
+	res, err := a.services.UrbsSetting.ProductDelete(ctx, product)
+	if err != nil {
+		return nil, err
+	}
+	err = a.daos.UrbsAcAcl.DeleteByObject(ctx, product)
 	if err != nil {
 		logger.Err(ctx, err.Error())
 	}
-	return a.services.UrbsSetting.ProductDelete(ctx, product)
+	return res, nil
 }
 
 // Statistics 返回产品的统计数据