+++ web/src/pages/components/Users.tsx
@@ -10,7 +10,16 @@ const Users: React.FC<TableComponentProps<any>> = (props) => {
     title: 'ID',
     dataIndex: 'user',
     key: 'user',
-  }, {
+  },
+  {
+    title: '创建时间',
+    dataIndex: 'createdAt',
+    key: 'createdAt',
+    render: (time: string) => {
+      return formatTableTime(time);
+    },
+  },
+  {
     title: '分配时间',
     dataIndex: 'assignedAt',
     key: 'assignedAt',